% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_power.r
\name{c_power}
\alias{c_power}
\title{Estimate power for conflict effect using monte-carlo simulation}
\usage{
c_power(subjects, c_nmst, nc_nmst, num_sims = 1000, alpha = 0.05)
}
\arguments{
\item{subjects}{A number for the number of subjects in simulated experiment}

\item{c_nmst}{A vector containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{nc_nmst}{A vector containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{num_sims}{A number, simulations to run}

\item{alpha}{A number, alpha criterion}
}
\value{
A number, power: the proportion of simulated experiments that returned a p-value less than the alpha criterion for the conflict effect
}
\description{
Estimate power for conflict effect using monte-carlo simulation
}
\details{
This function uses monte-carlo simulation to determine statistical power associated for detecting a conflict effect, and includes parameters for number of subjects in the experiment, number of trials in each condition (conflict vs. no-conflict), and parameters (mu,sigma,tau) for each reaction time distribution.

For every simulated experiment, a one sample t-test  (two-tailed) is computed, and the p-value is saved. Power is the proportion of simulated experiments that return p-values less than the defined alpha criterion.
}
\examples{
c_power(subjects=10,
  c_nmst = c(20,550,100,100),
  nc_nmst = c(20,500,100,100),
  num_sims = 1000,
  alpha = .05)
}
