% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_scores.r
\name{c_scores}
\alias{c_scores}
\title{Estimate conflict scores for n subjects}
\usage{
c_scores(
  subjects,
  c_nmst = c(n, mu, sigma, tau),
  nc_nmst = c(n, mu, sigma, tau)
)
}
\arguments{
\item{subjects}{A number for the number conflict scores to generate}

\item{c_nmst}{A vector containing the parameters for from an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{nc_nmst}{A vector containing the parameters for sampling from an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}
}
\value{
A simulated conflict score for one subject
}
\description{
Estimate conflict scores for n subjects
}
\details{
This function creates simulated conflict scores for n number of subjects. This is a wrapper for c_score, which creates a single conflict score.
}
\examples{
c_scores(subjects=50,
        c_nmst= c(n=10, mu=600, sigma=100, tau=100),
        nc_nmst= c(n=10, mu=500, sigma=100, tau=100))
}
