% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_modulation_power_fast.r
\name{pc_modulation_power_fast}
\alias{pc_modulation_power_fast}
\title{Estimate power for differences between proportion conflict effects using monte-carlo simulation}
\usage{
pc_modulation_power_fast(
  subjects,
  design,
  A_mc_c_nmst,
  A_mc_nc_nmst,
  A_mnc_c_nmst,
  A_mnc_nc_nmst,
  B_mc_c_nmst,
  B_mc_nc_nmst,
  B_mnc_c_nmst,
  B_mnc_nc_nmst,
  num_sims = 1000,
  alpha = 0.05
)
}
\arguments{
\item{subjects}{A number for the number of subjects in simulated experiment}

\item{design}{A string set to "paired" or "between"}

\item{A_mc_c_nmst}{A vector for mostly conflict (mc), conflict trials (c) containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{A_mc_nc_nmst}{A vector for mostly conflict (mc), conflict trials (c) containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{A_mnc_c_nmst}{A vector for mostly no-conflict (mnc), conflict trials (c) containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{A_mnc_nc_nmst}{A vector for mostly no-conflict (mnc), conflict trials (c) containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{B_mc_c_nmst}{A vector for mostly conflict (mc), conflict trials (c) containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{B_mc_nc_nmst}{A vector for mostly conflict (mc), conflict trials (c) containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{B_mnc_c_nmst}{A vector for mostly no-conflict (mnc), conflict trials (c) containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{B_mnc_nc_nmst}{A vector for mostly no-conflict (mnc), conflict trials (c) containing the parameters for an ex-gaussian distribution, c(n, mu, sigma, tau), where n is the number of trials.}

\item{num_sims}{A number, simulations to run}

\item{alpha}{A number, alpha criterion}
}
\value{
A number, power: the proportion of simulated experiments that returned a p-value less than the alpha criterion for the conflict effect
}
\description{
Estimate power for differences between proportion conflict effects using monte-carlo simulation
}
\details{
This function uses monte-carlo simulation to determine statistical power associated for detecting a modulation to a proportion conflict effect, specifically a difference between two proportion conflict effects.

Power is the proportion of simulated experiments that return p-values less than the defined alpha criterion.
}
\examples{
pc_modulation_power_fast(subjects=20,
  design ="paired",
  A_mc_c_nmst = c(80,550,100,100),
  A_mc_nc_nmst = c(20,500,100,100),
  A_mnc_c_nmst = c(20,570,100,100),
  A_mnc_nc_nmst = c(80,500,100,100),
  B_mc_c_nmst = c(80,530,100,100),
  B_mc_nc_nmst = c(20,500,100,100),
  B_mnc_c_nmst = c(20,570,100,100),
  B_mnc_nc_nmst = c(80,500,100,100),
  num_sims = 1000,
  alpha = .05)
}
